requirements.txt
fastapi
uvicorn
kafka-python
openai
openai-agents
requests
